SAR-6948 Integrated the submission builder with submission service, removed old formats
