+++ app/models/api/ApplicantDetails.scala
@@ -24,19 +24,17 @@ import play.api.libs.functional.syntax._
 import play.api.libs.json._
 import utils.JsonUtilities
 
-import scala.util.{Failure, Success, Try}
-
 case class ApplicantDetails(nino: String,
                             roleInBusiness: RoleInBusiness,
                             name: Name,
                             dateOfBirth: DateOfBirth,
                             companyName: String,
-                            companyNumber: Option[String] = None,
+                            companyNumber: String,
                             dateOfIncorporation: LocalDate,
-                            ctutr: Option[String] = None,
-                            businessVerification: Option[BusinessVerificationStatus] = None,
-                            registration: Option[BusinessRegistrationStatus] = None,
-                            identifiersMatch: Option[Boolean] = None,
+                            ctutr: String,
+                            businessVerification: BusinessVerificationStatus,
+                            registration: BusinessRegistrationStatus,
+                            identifiersMatch: Boolean,
                             bpSafeId: Option[String] = None,
                             currentAddress: Address,
                             contact: DigitalContactOptional,
@@ -48,22 +46,18 @@ object ApplicantDetails extends VatApplicantDetailsValidator
   with ApplicantDetailsHelper
   with JsonUtilities {
 
-  private val custInfoSection = JsPath \ "customerIdentification"
-  private val appDetailsSection = JsPath \ "declaration" \ "applicantDetails"
-  private val corpBodySection = JsPath \ "subscription" \ "corporateBodyRegistered"
-
   implicit val format: Format[ApplicantDetails] = (
     (__ \ "nino").format[String] and
       (__ \ "roleInTheBusiness").format[RoleInBusiness] and
       (__ \ "name").format[Name] and
       (__ \ "dateOfBirth").format[DateOfBirth] and
       (__ \ "companyName").format[String] and
-      (__ \ "companyNumber").formatNullable[String] and
+      (__ \ "companyNumber").format[String] and
       (__ \ "dateOfIncorporation").format[LocalDate] and
-      (__ \ "ctutr").formatNullable[String] and
-      (__ \ "businessVerification").formatNullable[BusinessVerificationStatus] and
-      (__ \ "registration").formatNullable[BusinessRegistrationStatus] and
-      (__ \ "identifiersMatch").formatNullable[Boolean] and
+      (__ \ "ctutr").format[String] and
+      (__ \ "businessVerification").format[BusinessVerificationStatus] and
+      (__ \ "registration").format[BusinessRegistrationStatus] and
+      (__ \ "identifiersMatch").format[Boolean] and
       (__ \ "bpSafeId").formatNullable[String] and
       (__ \ "currentAddress").format[Address] and
       (__ \ "contact").format[DigitalContactOptional] and
@@ -71,74 +65,4 @@ object ApplicantDetails extends VatApplicantDetailsValidator
       (__ \ "previousAddress").formatNullable[Address] and
       (__ \ "countryOfIncorporation").format[String]
     ) (ApplicantDetails.apply, unlift(ApplicantDetails.unapply))
-
-  val submissionReads: Reads[ApplicantDetails] = Reads[ApplicantDetails] { json =>
-    Try {
-      ApplicantDetails(
-        nino = json.nino,
-        roleInBusiness = json.getField[RoleInBusiness](appDetailsSection \ "roleInBusiness"),
-        name = json.getField[Name](appDetailsSection \ "name")(Name.submissionFormat),
-        dateOfBirth = json.getField[DateOfBirth](appDetailsSection \ "dateOfBirth"),
-        companyName = json.getField[String](custInfoSection \ "shortOrgName"),
-        companyNumber = json.getOptionalField[String](corpBodySection \ "companyRegistrationNumber").orElse(None),
-        dateOfIncorporation = json.getField[LocalDate](corpBodySection \ "dateOfIncorporation"),
-        ctutr = json.ctUtr.orElse(None),
-        businessVerification = json.businessVerificationStatus,
-        registration = json.businessRegistrationStatus,
-        identifiersMatch = json.identifiersMatch,
-        bpSafeId = json.getOptionalField[String](custInfoSection \ "primeBPSafeID").orElse(None),
-        currentAddress = json.getField[Address](appDetailsSection \ "currAddress")(Address.submissionFormat),
-        contact = json.getField[DigitalContactOptional](appDetailsSection \ "commDetails")(DigitalContactOptional.submissionFormat),
-        changeOfName = json.getOptionalField[FormerName](appDetailsSection \ "prevName")(FormerName.submissionFormat),
-        previousAddress = json.getOptionalField[Address](appDetailsSection \ "prevAddress")(Address.submissionFormat)
-      )
-    } match {
-      case Failure(exception) => JsError(exception.getMessage)
-      case Success(value) => JsSuccess(value)
-    }
-  }
-
-  val submissionWrites: Writes[ApplicantDetails] = Writes[ApplicantDetails] {
-    appDetails =>
-      Json.obj(
-        "customerIdentification" -> Json.obj(
-          "shortOrgName" -> appDetails.companyName
-        ).++(
-          optionalIds(appDetails)
-        ),
-        "subscription" -> Json.obj(
-          "corporateBodyRegistered" -> Json.obj(
-            "companyRegistrationNumber" -> appDetails.companyNumber,
-            "dateOfIncorporation" -> appDetails.dateOfIncorporation,
-            "countryOfIncorporation" -> "GB"
-          )
-        ),
-        "declaration" -> Json.obj(
-          "declarationSigning" -> Json.obj(
-            "declarationCapacity" -> Json.toJson(appDetails.roleInBusiness)
-          ),
-          "applicantDetails" -> Json.obj(
-            "roleInBusiness" -> Json.toJson(appDetails.roleInBusiness),
-            "name" -> Json.toJson(appDetails.name)(Name.submissionFormat),
-            "prevName" -> Json.toJson(appDetails.changeOfName.map(FormerName.submissionFormat.writes)),
-            "dateOfBirth" -> Json.toJson(appDetails.dateOfBirth),
-            "currAddress" -> Json.toJson(appDetails.currentAddress)(Address.submissionFormat),
-            "prevAddress" -> Json.toJson(appDetails.previousAddress.map(Address.submissionFormat.writes)),
-            "commDetails" -> Json.toJson(appDetails.contact)(DigitalContactOptional.submissionFormat),
-            "identifiers" -> Json.toJson(appDetails.personalIdentifiers)
-          ).filterNullFields
-        )
-      )
-  }
-
-  private def optionalIds(appDetails: ApplicantDetails): JsObject =
-    if (appDetails.bpSafeId.isDefined) {
-      Json.obj(
-        "primeBPSafeID" -> appDetails.bpSafeId.map(JsString)
-      )
-    }
-    else {
-      Json.obj("customerID" -> Json.toJson(appDetails.companyIdentifiers))
-    }
-
 }
