+++ app/models/api/BankAccount.scala
@@ -19,7 +19,6 @@ package models.api
 import auth.CryptoSCRS
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
-import uk.gov.hmrc.http.InternalServerException
 
 case class BankAccount(isProvided: Boolean,
                        details: Option[BankAccountDetails],
@@ -31,63 +30,10 @@ case class BankAccountDetails(name: String,
 
 object BankAccount {
   implicit val format: Format[BankAccount] = Json.format[BankAccount]
-
-  def submissionReads(optJson: Option[JsValue]): Option[BankAccount] = {
-    optJson match {
-      case Some(json) =>
-        val bankNameResult = (json \ "UK" \ "accountName").validate[String]
-        val sortCodeResult = (json \ "UK" \ "sortCode").validate[String]
-        val accNumberResult = (json \ "UK" \ "accountNumber").validate[String]
-        val reasonNoAccount = (json \ "UK" \ "reasonBankAccNotProvided").validate[NoUKBankAccount](NoUKBankAccount.submissionReads)
-
-        (bankNameResult, sortCodeResult, accNumberResult, reasonNoAccount) match {
-          case (JsSuccess(name, _), JsSuccess(sortCode, _), JsSuccess(accNumber, _), _) =>
-            Some(BankAccount(isProvided = true, details = Some(BankAccountDetails(name, sortCode, accNumber)), None))
-          case (_, _, _, JsSuccess(reason, _)) =>
-            Some(BankAccount(isProvided = false, details = None, reason = Some(reason)))
-          case _ =>
-            throw new Exception("Could not parse bank details")
-        }
-      case None =>
-        Some(BankAccount(isProvided = false, details = None, reason = None))
-    }
-
-  }
-
-  def submissionWrites: Writes[Option[BankAccount]] = Writes { optBankAccount: Option[BankAccount] =>
-    optBankAccount match {
-      case Some(BankAccount(true, Some(details), _)) =>
-        Json.obj("UK" -> Json.obj(
-          "accountName" -> details.name,
-          "sortCode" -> details.sortCode.replaceAll("-", ""),
-          "accountNumber" -> details.number
-        ))
-      case Some(BankAccount(false, _, Some(reason))) =>
-        Json.obj(
-          "UK" -> Json.obj(
-            "reasonBankAccNotProvided" -> Json.toJson(reason)(NoUKBankAccount.submissionWrites)
-          )
-        )
-      case _ => throw new InternalServerException("[BankAccount][submissionWrites] attempted to write with an invalid model")
-    }
-  }
 }
 
 object BankAccountDetails extends VatBankAccountValidator {
   implicit val format: Format[BankAccountDetails] = Json.format[BankAccountDetails]
-
-  val submissionFormat: OFormat[BankAccountDetails] = (
-    (__ \ "UK" \ "accountName").format[String] and
-      (__ \ "UK" \ "sortCode").format[String] and
-      (__ \ "UK" \ "accountNumber").format[String]
-    ) (BankAccountDetails.apply, unlift(BankAccountDetails.unapply))
-
-  val submissionReads: Reads[BankAccountDetails] = (
-    (__ \ "UK" \ "accountName").read[String] and
-      (__ \ "UK" \ "sortCode").read[String] and
-      (__ \ "UK" \ "accountNumber").read[String]
-    ) (BankAccountDetails.apply _)
-
 }
 
 object BankAccountDetailsMongoFormat extends VatBankAccountValidator {
