+++ conf/logback-test.xml
@@ -57,7 +57,7 @@
         <appender-ref ref="STDOUT"/>
     </logger>
 
-    <root level="OFF">
+    <root level="ERROR">
         <appender-ref ref="FILE"/>
         <appender-ref ref="STDOUT"/>
     </root>
