+++ test/mocks/VatMocks.scala
@@ -30,6 +30,7 @@ import org.scalatestplus.mockito.MockitoSugar
 import play.api.libs.json.JsValue
 import repositories._
 import services._
+import services.submission.SubmissionService
 import uk.gov.hmrc.auth.core.retrieve.{Retrieval, ~}
 import uk.gov.hmrc.auth.core.{AuthConnector, InvalidBearerToken}
 import uk.gov.hmrc.http.HeaderCarrier
