+++ test/services/submission/buildermocks/MockSubmissionPayloadBuilder.scala
@@ -22,18 +22,18 @@ import org.mockito.stubbing.OngoingStubbing
 import org.scalatest.Suite
 import org.scalatestplus.mockito.MockitoSugar
 import play.api.libs.json.JsObject
-import services.submission.ContactBlockBuilder
+import services.submission.SubmissionPayloadBuilder
 
 import scala.concurrent.Future
 
-trait MockContactBlockBuilder extends MockitoSugar {
+trait MockSubmissionPayloadBuilder extends MockitoSugar {
   self: Suite =>
 
-  val mockContactBlockBuilder: ContactBlockBuilder = mock[ContactBlockBuilder]
+  val mockSubmissionPayloadBuilder: SubmissionPayloadBuilder = mock[SubmissionPayloadBuilder]
 
-  def mockBuildContactBlock(regId: String)
-                           (response: Future[JsObject]): OngoingStubbing[Future[JsObject]] =
-    when(mockContactBlockBuilder.buildContactBlock(ArgumentMatchers.eq(regId)))
+  def mockBuildSubmissionPayload(regId: String)
+                                (response: Future[JsObject]): OngoingStubbing[Future[JsObject]] =
+    when(mockSubmissionPayloadBuilder.buildSubmissionPayload(ArgumentMatchers.eq(regId)))
       .thenReturn(response)
 
 }
